权限管理功能完善

+++ src/main/java/com/crud/demo/service/CourseService.java
@@ -1,7 +1,9 @@
 package com.crud.demo.service;
 
 import com.crud.demo.entity.StudentScore;
+import com.crud.demo.entity.Teacher;
 import com.crud.demo.mapper.StudentScoreMapper;
+import com.crud.demo.mapper.TeacherMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
@@ -17,15 +19,26 @@ public class CourseService{
     private CourseMapper courseMapper;
     @Autowired
     private StudentScoreMapper studentScoreMapper;
+    @Autowired
+    private TeacherMapper teacherMapper;
 
     
     public int deleteByPrimaryKey(Integer id) {
-        return courseMapper.deleteByPrimaryKey(id);
+        Course course = courseMapper.selectByPrimaryKey(id);
+        List<StudentScore> studentScoreList = studentScoreMapper.selectByCourseNumber(course.getCourseNumber());
+        if (studentScoreList.size() > 0){
+            return 0;
+        }else {
+            courseMapper.deleteByPrimaryKey(id);
+            return 1;
+        }
     }
 
     
-    public int insert(Course record) {
-        return courseMapper.insert(record);
+    public int insert(Course course) {
+        Teacher teacher = teacherMapper.getTeacherByNumber(course.getTNumber());
+        course.setTName(teacher.getTName());
+        return courseMapper.insert(course);
     }
 
     
@@ -49,7 +62,12 @@ public class CourseService{
     }
 
     public void updateCourseAndScoreById(Course course){
+        //根据教师号填充教师名
+        Teacher teacher = teacherMapper.getTeacherByNumber(course.getTNumber());
+        course.setTName(teacher.getTName());
+
         courseMapper.updateByPrimaryKey(course);
+
         Course dbCourse = courseMapper.selectByPrimaryKey(course.getId());
         List<StudentScore> studentScoreList = studentScoreMapper.selectByCourseNumber(dbCourse.getCourseNumber());
         //遍历成绩信息，更改课程名
